use BigDecimal instead of Double in IS24-CSV-Format

+++ IS24-CSV/src/main/java/org/openestate/io/is24_csv/records/GewerbeEinzelhandel.java
@@ -16,6 +16,7 @@
 
 package org.openestate.io.is24_csv.records;
 
+import java.math.BigDecimal;
 import org.apache.commons.csv.CSVRecord;
 import org.openestate.io.is24_csv.Is24CsvFormat;
 import org.openestate.io.is24_csv.Is24CsvRecord;
@@ -215,11 +216,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
       this.get( FIELD_BODENBELAG ) );
   }
 
-  public Double getDeckenlast()
+  public BigDecimal getDeckenlast()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_DECKENLAST ) );
     }
     catch (NumberFormatException ex)
@@ -242,11 +243,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
       this.get( FIELD_ENERGIEAUSWEIS_INKL_WARMWASSER ) );
   }
 
-  public Double getEnergieausweisKennwert()
+  public BigDecimal getEnergieausweisKennwert()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_ENERGIEAUSWEIS_KENNWERT ) );
     }
     catch (NumberFormatException ex)
@@ -328,11 +329,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
     }
   }
 
-  public Double getGesamtflaeche()
+  public BigDecimal getGesamtflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_GESAMTFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -343,11 +344,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
     }
   }
 
-  public Double getGewerbeflaeche()
+  public BigDecimal getGewerbeflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_GEWERBEFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -364,11 +365,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
       this.get( FIELD_HEIZUNGSART ) );
   }
 
-  public Double getKaltmiete()
+  public BigDecimal getKaltmiete()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_KALTMIETE ) );
     }
     catch (NumberFormatException ex)
@@ -385,11 +386,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
       this.get( FIELD_KALTMIETE_PRO ) );
   }
 
-  public Double getKaufpreis()
+  public BigDecimal getKaufpreis()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_KAUFPREIS ) );
     }
     catch (NumberFormatException ex)
@@ -423,11 +424,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
       this.get( FIELD_LASTENAUFZUG ) );
   }
 
-  public Double getLastenaufzugTragkraft()
+  public BigDecimal getLastenaufzugTragkraft()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_LASTENAUFZUG_TRAGKRAFT ) );
     }
     catch (NumberFormatException ex)
@@ -438,11 +439,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
     }
   }
 
-  public Double getNebenflaeche()
+  public BigDecimal getNebenflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_NEBENFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -453,11 +454,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
     }
   }
 
-  public Double getNebenkosten()
+  public BigDecimal getNebenkosten()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_NEBENKOSTEN ) );
     }
     catch (NumberFormatException ex)
@@ -486,11 +487,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
       this.get( FIELD_PERSONENAUFZUG ) );
   }
 
-  public Double getPreisProParkflaeche()
+  public BigDecimal getPreisProParkflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_PREIS_PRO_PARKFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -522,11 +523,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
     }
   }
 
-  public Double getSchaufensterfront()
+  public BigDecimal getSchaufensterfront()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_SCHAUFENSTERFRONT ) );
     }
     catch (NumberFormatException ex)
@@ -537,11 +538,11 @@ public class GewerbeEinzelhandel extends Is24CsvRecord
     }
   }
 
-  public Double getTeilbarAb()
+  public BigDecimal getTeilbarAb()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_TEILBAR_AB ) );
     }
     catch (NumberFormatException ex)
