+++ IS24-CSV/src/main/java/org/openestate/io/is24_csv/records/GewerbeHalleProduktion.java
@@ -16,6 +16,7 @@
 
 package org.openestate.io.is24_csv.records;
 
+import java.math.BigDecimal;
 import org.apache.commons.csv.CSVRecord;
 import org.openestate.io.is24_csv.Is24CsvFormat;
 import org.openestate.io.is24_csv.Is24CsvRecord;
@@ -206,11 +207,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_BODENBELAG ) );
   }
 
-  public Double getBodenbelastung()
+  public BigDecimal getBodenbelastung()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_BODENBELASTUNG ) );
     }
     catch (NumberFormatException ex)
@@ -227,11 +228,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_ENERGIEAUSWEIS_INKL_WARMWASSER ) );
   }
 
-  public Double getEnergieausweisKennwert()
+  public BigDecimal getEnergieausweisKennwert()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_ENERGIEAUSWEIS_KENNWERT ) );
     }
     catch (NumberFormatException ex)
@@ -313,11 +314,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getGesamtflaeche()
+  public BigDecimal getGesamtflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_GESAMTFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -328,11 +329,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getGewerbeflaeche()
+  public BigDecimal getGewerbeflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_GEWERBEFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -343,11 +344,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getGrundstuecksflaeche()
+  public BigDecimal getGrundstuecksflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_GRUNDSTUECKSFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -358,11 +359,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getHallenhoehe()
+  public BigDecimal getHallenhoehe()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_HALLENHOEHE ) );
     }
     catch (NumberFormatException ex)
@@ -385,11 +386,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_HEIZUNGSART ) );
   }
 
-  public Double getKaltmiete()
+  public BigDecimal getKaltmiete()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_KALTMIETE ) );
     }
     catch (NumberFormatException ex)
@@ -406,11 +407,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_KALTMIETE_PRO ) );
   }
 
-  public Double getKaufpreis()
+  public BigDecimal getKaufpreis()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_KAUFPREIS ) );
     }
     catch (NumberFormatException ex)
@@ -432,11 +433,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_KRANBAHN ) );
   }
 
-  public Double getKranbahnTragkraft()
+  public BigDecimal getKranbahnTragkraft()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_KRANBAHN_TRAGKRAFT ) );
     }
     catch (NumberFormatException ex)
@@ -453,11 +454,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_LASTENAUFZUG ) );
   }
 
-  public Double getLastenaufzugTragkraft()
+  public BigDecimal getLastenaufzugTragkraft()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_LASTENAUFZUG_TRAGKRAFT ) );
     }
     catch (NumberFormatException ex)
@@ -468,11 +469,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getNebenflaeche()
+  public BigDecimal getNebenflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_NEBENFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -483,11 +484,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getNebenkosten()
+  public BigDecimal getNebenkosten()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_NEBENKOSTEN ) );
     }
     catch (NumberFormatException ex)
@@ -516,11 +517,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
       this.get( FIELD_PERSONENAUFZUG ) );
   }
 
-  public Double getPreisProParkflaeche()
+  public BigDecimal getPreisProParkflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_PREIS_PRO_PARKFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -567,11 +568,11 @@ public class GewerbeHalleProduktion extends Is24CsvRecord
     }
   }
 
-  public Double getTeilbarAb()
+  public BigDecimal getTeilbarAb()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_TEILBAR_AB ) );
     }
     catch (NumberFormatException ex)
