+++ IS24-CSV/src/main/java/org/openestate/io/is24_csv/records/WohnungMiete.java
@@ -16,6 +16,7 @@
 
 package org.openestate.io.is24_csv.records;
 
+import java.math.BigDecimal;
 import org.apache.commons.csv.CSVRecord;
 import org.openestate.io.is24_csv.Is24CsvFormat;
 import org.openestate.io.is24_csv.Is24CsvRecord;
@@ -263,11 +264,11 @@ public class WohnungMiete extends Is24CsvRecord
       this.get( FIELD_ENERGIEAUSWEIS_INKL_WARMWASSER ) );
   }
 
-  public Double getEnergieausweisKennwert()
+  public BigDecimal getEnergieausweisKennwert()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_ENERGIEAUSWEIS_KENNWERT ) );
     }
     catch (NumberFormatException ex)
@@ -338,11 +339,11 @@ public class WohnungMiete extends Is24CsvRecord
       this.get( FIELD_HAUSTIERE ) );
   }
 
-  public Double getHeizkosten()
+  public BigDecimal getHeizkosten()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_HEIZKOSTEN ) );
     }
     catch (NumberFormatException ex)
@@ -359,11 +360,11 @@ public class WohnungMiete extends Is24CsvRecord
       this.get( FIELD_HEIZUNGSART ) );
   }
 
-  public Double getKaltmiete()
+  public BigDecimal getKaltmiete()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_KALTMIETE ) );
     }
     catch (NumberFormatException ex)
@@ -385,11 +386,11 @@ public class WohnungMiete extends Is24CsvRecord
       this.get( FIELD_KELLER ) );
   }
 
-  public Double getNebenkosten()
+  public BigDecimal getNebenkosten()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_NEBENKOSTEN ) );
     }
     catch (NumberFormatException ex)
@@ -406,11 +407,11 @@ public class WohnungMiete extends Is24CsvRecord
       this.get( FIELD_NEBENKOSTEN_INKL_HEIZKOSTEN ) );
   }
 
-  public Double getNutzflaeche()
+  public BigDecimal getNutzflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_NUTZFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -467,11 +468,11 @@ public class WohnungMiete extends Is24CsvRecord
       this.get( FIELD_STELLPLATZ ) );
   }
 
-  public Double getStellplatzmiete()
+  public BigDecimal getStellplatzmiete()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_STELLPLATZMIETE ) );
     }
     catch (NumberFormatException ex)
@@ -487,11 +488,11 @@ public class WohnungMiete extends Is24CsvRecord
     return this.get( FIELD_VERFUEGBAR_AB );
   }
 
-  public Double getWarmmiete()
+  public BigDecimal getWarmmiete()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_WARMMIETE ) );
     }
     catch (NumberFormatException ex)
@@ -502,11 +503,11 @@ public class WohnungMiete extends Is24CsvRecord
     }
   }
 
-  public Double getWohnflaeche()
+  public BigDecimal getWohnflaeche()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_WOHNFLAECHE ) );
     }
     catch (NumberFormatException ex)
@@ -517,11 +518,11 @@ public class WohnungMiete extends Is24CsvRecord
     }
   }
 
-  public Double getZimmer()
+  public BigDecimal getZimmer()
   {
     try
     {
-      return Is24CsvFormat.parseDouble(
+      return Is24CsvFormat.parseDecimal(
         this.get( FIELD_ZIMMER ) );
     }
     catch (NumberFormatException ex)
